ProcChain -> ProcNetwork, NetworkHandler -> NetworkChain
The ProcNetwork(InOut) functions are designed to help you spin up a
process network and return the external channels to the caller. Each
component given will have a simple accept/execute loop and then based on
the result of the component handler will pass the connection on to the
next component (without knowing what it is).

The NetworkChain function has been introduced to take an input channel
and output channel and use them exclusively to distribute requests to
the process network. Splitting the logic between these two handlers has
made things much cleaner and more powerful at the same time.